requirements.txt
flask==3.1.0
gunicorn==23.0.0
pytest==7.4.2
pytest-cov==4.1.0
requests==2.32.3
flask==2.2.3
psycopg2-binary==2.9.5
python-dotenv==1.0.0